requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pyTelegramBotAPI==4.22.0
redis==5.0.8
requests==2.32.3
telebot-calendar==1.2
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
